implemented until obtain bookings; TODO() -> fix reading a full line for makeBooking        -> update column estado from bicicleta when making or cancelling a booking

+++ model/src/main/java/isel/sisinf/model/ReservaInfo.java
@@ -1,13 +1,15 @@
 package isel.sisinf.model;
 
+import java.math.BigDecimal;
+
 public class ReservaInfo {
     private Integer numero;
     private String dtInicio;
     private String dtFim;
-    private Double valor;
+    private BigDecimal valor;
     private Integer bicicletaId;
 
-    public ReservaInfo(Integer numero, String dtInicio, String dtFim, Double valor, Integer bicicletaId) {
+    public ReservaInfo(Integer numero, String dtInicio, String dtFim, BigDecimal valor, Integer bicicletaId) {
         this.numero = numero;
         this.dtInicio = dtInicio;
         this.dtFim = dtFim;
